+++ FantasyStats/DataManager/Models/CalculatedPlayerStatistics.cs
@@ -14,7 +14,7 @@ namespace DataManager.Models
         public double xPAbs { get; set; }
         public double xPPound90 { get; set; }
         public double xPPoundMinPlayed { get; set; }
-        public double MinutesPlayed { get; set; }
+        public int MinutesPlayed { get; set; }
         public double xYc { get; set; }
         public double xRc { get; set; }
     }
